Smart feature additions

+++ unifood/lib/view_model/plate_view_model.dart
@@ -15,7 +15,7 @@ class PlateViewModel {
               imagePath: item['imageURL'],
               name: item['name'],
               description: item['description'],
-              price: item['price'],
+              price: item['price'].toDouble(),
             ),
           )
           .toList();
